Rename animations (#4)
* clean up
