+++ animations/animation_original.py
@@ -1,14 +1,17 @@
-from rgbmatrix import graphics
 import time
 
 from leezenflow_base import LeezenflowBase
-from shared_state import SharedState
 
-class AnimationOrignal(LeezenflowBase):
-    def __init__(self, command_line_args):
-        print("a")
-        super(AnimationOrignal, self).__init__(command_line_args)
-        print("c")
+from shared_state import SharedState
+from command_line_args import CommandLineArgs
+
+class AnimationOriginal(LeezenflowBase):
+    
+    def __init__(self, command_line_args, display):
+        super(AnimationOriginal, self).__init__(command_line_args, display)
+
+        self.display = display
+
         self.bike_height = None
         self.bike1_position = None
         self.bike2_position = None
@@ -27,85 +30,72 @@ class AnimationOrignal(LeezenflowBase):
 
         self.color_green = (0,255,132)
         self.color_red = (255,49,73)
-        self.length_shade = 25
-        self.bias = 3
         self.placeholder_time_for_short_phases = 45
 
         self.last_update_current_row = None
         self.last_update_prediction_sec = None
         self.last_update_elapsed_sec = None
         self.last_update_elapsed_time_start = None
-       
+
+        self.lastGreenPhaseTime = -1
+        self.averageGreenPhaseTime = -1
+        self.greenPhaseArray = []
+        self.greenPhasesToSave = 10
+        self.lastRedPhaseTime = -1
+        self.averageRedPhaseTime = -1
+        self.redPhaseArray = []
+        self.redPhasesToSave = 10
+        self.firstRun = True
+        self.switchedPhase = False
+        self.starttime = time.time()
+        self.seconds_to_reach_traffic_light = 0
+        self.bicycle_speed = 0 # in m/s
+        self.distance = 0
+        self.first_initialize = True
+
     def run(self,_,run_event):
 
         g1,g2,g3 = *self.color_green,
         r1,r2,r3 = *self.color_red,
-        canvas = self.matrix
 
-        def setRow(row_nr):
-            for pixel in range(0,32):
-                self.matrix.SetPixel(row_nr, pixel, 0, 0, 0)
 
-        def draw_rectangle(y1,y2,r,g,b):
-            for row in range(y1,y2):
-                for pixel in range(0,32):
-                    self.matrix.SetPixel(row, pixel, r, g, b)
 
-        def draw_rectangle_shade(y1,y2,r,g,b):
-            shade_intensity = self.length_shade *2
-            for row in range(y1,y2):
-                for pixel in range(0,32):
-                    self.matrix.SetPixel(row, pixel, r/shade_intensity, g/shade_intensity, b/shade_intensity)
+        def calculate_seconds_to_reach_traffic_light():
+            # time = distance (in meter) / speed (in m/s)
+            self.bicycle_speed = CommandLineArgs.command_line_args.bicycle_speed / 3.6 # calculate km/h to m/s
+            self.distance = CommandLineArgs.command_line_args.distance
+            if (self.distance > 0 and self.bicycle_speed > 0):
+                self.seconds_to_reach_traffic_light = self.distance / self.bicycle_speed
+                print("seconds_to_reach_traffic_light " , self.seconds_to_reach_traffic_light)
 
-        def draw_black_trend_rectangle(y1,y2,r,g,b):
-            y1 -= self.bias
-            y2 -= self.bias
-            length = self.length_shade # How many shaded pixel?
-            r1 = r/length
-            g1 = g/length
-            b1 = b/length
+        calculate_seconds_to_reach_traffic_light()
 
-            for row in range(0,length): # Fade of length 25 upwards starting at y2
-                for pixel in range(0,32):
-                    self.matrix.SetPixel(y2+row, pixel, r1*(row+1), g1*(row+1), b1*(row+1))
-
-            for row in range(y1,y2): # Plain dark green/red at the bottom
-                for pixel in range(0,32):
-                    self.matrix.SetPixel(row, pixel, r1/2, g1/2, b1/2) # Dark green/red
-
-        # Function to manually draw a bike. It should also be possible (regarding performance) to include a .png graphic instead.
-        def draw_bike(color,y_position,axis_x_left,moving=False):
-            axis_x_left = axis_x_left
-            axis_y = y_position
-            axis_x_middle = axis_x_left + 7
-            axis_x_right = axis_x_left + 15
-            radius = 5
-
-            # Wheels
-            graphics.DrawCircle(canvas, axis_y, axis_x_left, radius, color)
-            graphics.DrawCircle(canvas, axis_y, axis_x_right, radius, color)
-
-            if moving:    
-                # Speed lines
-                graphics.DrawLine(canvas, axis_y-radius+1, axis_x_left-radius-7, axis_y-radius+1, axis_x_left-radius-1, color)
-                graphics.DrawLine(canvas, axis_y-radius+4, axis_x_left-radius-4, axis_y-radius+4, axis_x_left-radius-2, color)
-                graphics.DrawLine(canvas, axis_y-radius+8, axis_x_left-radius-6, axis_y-radius+8, axis_x_left-radius-2, color)
-
-            graphics.DrawLine(canvas, axis_y, axis_x_left, axis_y, axis_x_middle, color)
-            graphics.DrawLine(canvas, axis_y, axis_x_left, axis_y+8, axis_x_left+4, color)
-            graphics.DrawLine(canvas, axis_y, axis_x_middle, axis_y+8,  axis_x_middle+4, color)
-            graphics.DrawLine(canvas, axis_y+8, axis_x_left+4, axis_y+8,  axis_x_middle+4, color)
-            graphics.DrawLine(canvas, axis_y, axis_x_right, axis_y+8, axis_x_right-4, color)
-            graphics.DrawLine(canvas, axis_y+8,  axis_x_middle+4, axis_y+10,  axis_x_middle+5, color)
-            graphics.DrawLine(canvas, axis_y+10,  axis_x_middle+5, axis_y+10,  axis_x_middle+7, color)
-            graphics.DrawLine(canvas, axis_y+9, axis_x_left+4, axis_y+10,  axis_x_left+3, color)
-            graphics.DrawLine(canvas, axis_y+11,  axis_x_left+1, axis_y+11,  axis_x_left+5, color)
-          
         def update_current_row():
             if self.last_update_prediction_sec <= 0.1:
-                print("Prediction <= 0.1", flush=True)
+                #print("Prediction <= 0.1", flush=True)
                 return self.matrix_height
             self.current_row = self.last_update_current_row + int(((self.matrix_height - self.last_update_current_row) / self.last_update_prediction_sec) * self.last_update_elapsed_sec)
+
+        def process_predict_update_early_switch(phase):
+            self.last_update_elapsed_time_start = time.monotonic()
+
+            self.current_row = self.bike_box_height
+            self.last_update_current_row = self.current_row
+            self.last_update_elapsed_sec = 0
+            average_phase_time = 0
+
+            if phase == "red":
+                if len(self.redPhaseArray) > 3:
+                    average_phase_time = self.averageRedPhaseTime
+                else:
+                    average_phase_time = 50
+            else:
+                if len(self.greenPhaseArray) > 3:
+                    average_phase_time = self.averageGreenPhaseTime
+                else:
+                    average_phase_time = 50
+
+            self.last_update_prediction_sec = max(0, average_phase_time + self.seconds_to_reach_traffic_light)
 
         def process_prediction_update():
             self.last_update_current_row = self.current_row
@@ -113,7 +103,7 @@ class AnimationOrignal(LeezenflowBase):
             if SharedState.shared_data["current_phase"] == "yellow" or SharedState.shared_data["current_phase"] == "red-yellow":
                 self.last_update_prediction_sec = self.placeholder_time_for_short_phases
             else:
-                self.last_update_prediction_sec = max(0, SharedState.shared_data["change_timestamp"] - SharedState.shared_data["current_timestamp"])
+                self.last_update_prediction_sec = max(0, (SharedState.shared_data["change_timestamp"] - SharedState.shared_data["current_timestamp"]) - self.seconds_to_reach_traffic_light)
             self.last_update_elapsed_time_start = time.monotonic()
 
         def get_next_bike_step(time_elapsed_bike,last_bike_time):
@@ -122,20 +112,20 @@ class AnimationOrignal(LeezenflowBase):
                 return True
 
         def move_bikes_green():
-            draw_bike(graphics.Color(self.max_white, self.max_white, self.max_white),self.bike_height,self.bike1_position,moving=True)
+            self.display.draw_bike(self.max_white, self.max_white, self.max_white ,self.bike_height,self.bike1_position,moving=True)
             if get_next_bike_step(self.time_elapsed_bike,self.last_bike_time): 
                 self.bike1_position += 1 
-                self.bike2_position += 1           
+                self.bike2_position += 1
             if self.bike1_position >= 37: 
                 self.bike1_position = -20
                 self.bike2_position = 37
             if self.bike2_position >= 37:
-                draw_bike(graphics.Color(self.max_white, self.max_white, self.max_white),self.bike_height,self.bike2_position,moving=True)
+                self.display.draw_bike(self.max_white, self.max_white, self.max_white ,self.bike_height,self.bike2_position,moving=True)
                 if self.bike2_position >= 45:
-                    self.bike2_position = -999       
+                    self.bike2_position = -999
 
         def move_bike_red():
-            draw_bike(graphics.Color(self.max_white, self.max_white, self.max_white),self.bike_height,self.bike1_position,moving=True)
+            self.display.draw_bike(self.max_white, self.max_white, self.max_white ,self.bike_height,self.bike1_position,moving=True)
             if get_next_bike_step(self.time_elapsed_bike,self.last_bike_time): 
                 self.bike1_position += 1
             if self.bike1_position >= 45: # If the bike's position is beyond the center, the bike leaves the matrix and respawns in the center
@@ -146,61 +136,105 @@ class AnimationOrignal(LeezenflowBase):
 
         def initialize_phase_variables():
             # Time for rows
-            self.last_update_elapsed_time_start = time.monotonic()   
+            self.last_update_elapsed_time_start = time.monotonic()
             self.last_update_elapsed_time = 0
 
             # Time for bikes
-            self.start_time_bike = time.monotonic()            
+            self.start_time_bike = time.monotonic()
             self.time_elapsed_bike = 0
 
-            self.last_update_prediction_sec = max(0, SharedState.shared_data["change_timestamp"] - SharedState.shared_data["current_timestamp"])
+            self.last_update_prediction_sec = max(0, (SharedState.shared_data["change_timestamp"] - SharedState.shared_data["current_timestamp"]) - self.seconds_to_reach_traffic_light)
             self.last_message = None
-            self.current_row = self.bike_box_height
+            if self.first_initialize == True:
+                self.current_row = self.bike_box_height
+                self.first_initialize = False
             self.last_update_current_row = self.current_row
 
         ### Start the loop ###
         while(run_event.is_set()):
+
             ## Green phase ##
             if SharedState.shared_data["current_phase"] == "green" or SharedState.shared_data["current_phase"] == "red-yellow":
-              
+
                 # There are two bikes such that the second bike can roll into the matrix before the other has left 
-                self.last_bike_time = 0.0                
+                self.last_bike_time = 0.0
                 self.bike1_position = 8 
                 self.bike2_position = -999
                 self.bike_still_moving = True
 
-                initialize_phase_variables()
-                draw_rectangle(self.bike_box_height,self.matrix_height,g1,g2,g3) # pure green
+                self.switchedPhase = False
 
+                initialize_phase_variables()
+                if self.firstRun == True:
+                    self.firstRun = False
+                    self.display.draw_rectangle(self.bike_box_height,self.matrix_height, g1, g2, g3) # pure green
+    
                 while(self.current_row < self.matrix_height) and run_event.is_set():
+
                     self.time_elapsed_bike = time.monotonic() - self.start_time_bike
                     self.last_update_elapsed_sec = time.monotonic() - self.last_update_elapsed_time_start
 
                     update_current_row()
-                    draw_black_trend_rectangle(0,self.current_row,g1,g2,g3) # Black fade and resetting of bike box
+                    self.display.draw_black_trend_rectangle(0,self.current_row, g1, g2, g3) # Black fade and resetting of bike box
                     move_bikes_green()
 
                     if SharedState.shared_data["hash"] != self.last_message: # This avoids updates with every message that would break animation. Problematic with additional info in shared data.    
                         if SharedState.shared_data["current_phase"] != "green" and SharedState.shared_data["current_phase"] != "red-yellow":
                             break
                         process_prediction_update()
-                        draw_rectangle(self.bike_box_height+self.current_row,self.matrix_height,g1,g2,g3) # pure green
+                        self.display.draw_rectangle(self.bike_box_height+self.current_row,self.matrix_height, g1, g2, g3) # pure green
                         self.last_message = SharedState.shared_data["hash"]
 
+                        if SharedState.shared_data["current_phase"] == "green":
+                            self.lastGreenPhaseTime = max(0, SharedState.shared_data["change_timestamp"] - SharedState.shared_data["current_timestamp"])
+                            if len(self.greenPhaseArray) >= self.greenPhasesToSave:
+                                self.greenPhaseArray.pop(0)
+                            self.greenPhaseArray.append(self.lastGreenPhaseTime)
+                            self.averageGreenPhaseTime = 0
+                            for x in self.greenPhaseArray:
+                                self.averageGreenPhaseTime += x
+                            self.averageGreenPhaseTime = self.averageGreenPhaseTime / len(self.greenPhaseArray)
+
                     time.sleep(self.update_interval)
+
+                self.switchedPhase = True
+                just_switched = True
+
+                while SharedState.shared_data["current_phase"] == "green" and (self.current_row >= self.matrix_height) and run_event.is_set():
+                    self.current_row = self.bike_box_height
+                    self.display.draw_rectangle(self.bike_box_height,self.matrix_height, r1, r2, r3) # pure red
+                    process_predict_update_early_switch("green")
+
+                    while SharedState.shared_data["current_phase"] == "green":
+                        self.time_elapsed_bike = time.monotonic() - self.start_time_bike
+                        self.last_update_elapsed_sec = time.monotonic() - self.last_update_elapsed_time_start
+
+
+                        update_current_row()
+                        self.display.draw_black_trend_rectangle(0,self.current_row, r1, r2, r3) # Black fade and resetting of bike box
+
+                        if self.bike_still_moving:
+                            self.display.draw_black_trend_rectangle(0,self.current_row, r1, r2, r3) # Black fade
+                            move_bike_red()
+                        else:
+                            self.display.draw_bike(self.max_white, self.max_white, self.max_white ,self.bike_height,self.bike_center)
+
+                        
+
+                        time.sleep(self.update_interval)
 
                 # Wait until there is a message that the phase has actually changed
                 time_wait = time.monotonic()
                 while SharedState.shared_data["current_phase"] == "green" and run_event.is_set():
                     # Continue to move bikes
-                    draw_rectangle_shade(0,self.matrix_height,g1,g2,g3) # Just to reset each bike frame 
+                    self.display.draw_rectangle_shade(0,self.matrix_height, g1, g2, g3) # Just to reset each bike frame 
                     move_bikes_green()
                     self.time_elapsed_bike = time.monotonic() - self.start_time_bike
 
                     time.sleep(self.update_interval)
                     if time.monotonic() >= time_wait + 60: # 60 sec w/o expected phase change
-                        print("Green phase prediction time ended 60s ago, however no red phase message was received. Leaving green state...",flush=True)
-                        self.matrix.Fill(0,0,0)
+                        #print("Green phase prediction time ended 60s ago, however no red phase message was received. Leaving green state...",flush=True)
+                        self.display.Fill(0,0,0)
                         SharedState.shared_data["current_phase"] = "awaiting_message"
                         break
 
@@ -210,10 +244,15 @@ class AnimationOrignal(LeezenflowBase):
                 initialize_phase_variables()
                 just_switched = True
 
+                self.switchedPhase = False
+
                 # Red phase while bike moving. The bike moves beyond the green phase until it is in centered position.
                 self.last_bike_time = 0.0
-                draw_rectangle(self.bike_box_height,self.matrix_height,r1,r2,r3) # Red               
-                
+
+                if self.firstRun == True:
+                    self.firstRun = False
+                    self.display.draw_rectangle(self.bike_box_height,self.matrix_height, r1, r2, r3) # Red
+                    
                 while((self.current_row < self.matrix_height) and run_event.is_set()):
                     self.time_elapsed_bike = time.monotonic() - self.start_time_bike
                     self.last_update_elapsed_sec = time.monotonic() - self.last_update_elapsed_time_start
@@ -221,31 +260,58 @@ class AnimationOrignal(LeezenflowBase):
                     update_current_row()
 
                     if self.bike_still_moving:
-                        draw_black_trend_rectangle(0,self.current_row,r1,r2,r3) # Black fade
+                        self.display.draw_black_trend_rectangle(0,self.current_row, r1, r2, r3) # Black fade
                         move_bike_red()
                     else:
                         if just_switched:
-                            draw_rectangle_shade(0,self.bike_box_height,r1,r2,r3) # Reset bottom bike box to shaded red
-                            draw_bike(graphics.Color(self.max_white, self.max_white, self.max_white),self.bike_height,self.bike_center)
+                            self.display.draw_rectangle_shade(0,self.bike_box_height, r1, r2, r3) # Reset bottom bike box to shaded red
+                            self.display.draw_bike(self.max_white, self.max_white, self.max_white ,self.bike_height,self.bike_center)
                             just_switched = False
-                        draw_black_trend_rectangle(self.bike_box_height,self.current_row,r1,r2,r3) # Black fade
+                        self.display.draw_black_trend_rectangle(self.bike_box_height,self.current_row, r1, r2, r3) # Black fade
 
                     if SharedState.shared_data["hash"] != self.last_message: # This avoids updates with every message that would break animation. Problematic with additional info in shared data.
                         if SharedState.shared_data["current_phase"] != "red" and SharedState.shared_data["current_phase"] !="yellow":
                             break
                         process_prediction_update()
-                        draw_rectangle(self.bike_box_height+self.current_row,self.matrix_height,r1,r2,r3) # pure red 
-                        self.last_message = SharedState.shared_data["hash"] 
+                        self.display.draw_rectangle(self.bike_box_height+self.current_row,self.matrix_height, r1, r2, r3) # pure red
+                        self.last_message = SharedState.shared_data["hash"]
+
+                        if SharedState.shared_data["current_phase"] == "red":
+                            self.lastRedPhaseTime = max(0, SharedState.shared_data["change_timestamp"] - SharedState.shared_data["current_timestamp"])
+                            if len(self.redPhaseArray) >= self.redPhasesToSave:
+                                self.redPhaseArray.pop(0)
+                            self.redPhaseArray.append(self.lastRedPhaseTime)
+                            self.averageRedPhaseTime = 0
+                            for x in self.redPhaseArray:
+                                self.averageRedPhaseTime += x
+                            self.averageRedPhaseTime = self.averageRedPhaseTime / len(self.redPhaseArray)
 
                     time.sleep(self.update_interval)
+
+                self.switchedPhase = True
+                while SharedState.shared_data["current_phase"] == "red" and (self.current_row >= self.matrix_height) and run_event.is_set():
+                    self.current_row = self.bike_box_height
+                    self.display.draw_rectangle(self.bike_box_height,self.matrix_height, g1, g2, g3) # pure green
+                    process_predict_update_early_switch("green")
+
+                    while SharedState.shared_data["current_phase"] == "red":
+                        self.time_elapsed_bike = time.monotonic() - self.start_time_bike
+                        self.last_update_elapsed_sec = time.monotonic() - self.last_update_elapsed_time_start
+                        self.display.draw_black_trend_rectangle(0,self.current_row, g1, g2, g3) # Black fade and resetting of bike box
+                        move_bikes_green()
+                        update_current_row()
+                        
+                        time.sleep(self.update_interval)
+
+
 
                 # Wait until there is a message that the phase has actually changed
                 time_wait = time.monotonic()
                 while SharedState.shared_data["current_phase"] == "red" and run_event.is_set():
                     time.sleep(0.1)
                     if time.monotonic() >= time_wait + 60: # 60 sec w/o expected phase change
-                        print("Red phase prediction time ended 10s ago, however no green phase message was received. Leaving red state...",flush=True)
-                        self.matrix.Fill(0,0,0)
+                        #print("Red phase prediction time ended 10s ago, however no green phase message was received. Leaving red state...",flush=True)
+                        self.display.Fill(0,0,0)
                         SharedState.shared_data["current_phase"] = "awaiting_message"
                         break
 
@@ -256,7 +322,7 @@ class AnimationOrignal(LeezenflowBase):
                 down = True
                 while SharedState.shared_data["current_phase"] == "awaiting_message" and run_event.is_set():
                     time.sleep(0.1)
-                    draw_bike(graphics.Color(pulse, pulse, pulse),self.bike_height,self.bike_center)
+                    self.display.draw_bike(pulse, pulse, pulse, self.bike_height, self.bike_center)
                     if down:
                         pulse -= 10
                         if pulse <= 0:
@@ -268,13 +334,13 @@ class AnimationOrignal(LeezenflowBase):
                             pulse = 255
                             down = True
                     if time.monotonic() >= time_wait + 600: # After 10 min
-                        draw_rectangle(0,self.matrix_height,0,0,0)
-                        self.matrix.SetPixel(1, 1, 10, 10, 10) # Debug pixel
-                        while SharedState.shared_data["current_phase"] == "awaiting_message" and run_event.is_set():                                                      
+                        self.display.draw_rectangle(0,self.matrix_height,0,0,0)
+                        self.display.SetPixel(1, 1, 10, 10, 10)
+                        while SharedState.shared_data["current_phase"] == "awaiting_message" and run_event.is_set():
                             time.sleep(1)
             else:
                 print("Unknown phase / error: ",SharedState.shared_data,flush=True)
-                draw_bike(graphics.Color(255,49,73),self.bike_height,self.bike_center)
+                self.display.draw_bike(255, 49, 73, self.bike_height, self.bike_center)
                 time.sleep(15)
                 print("Trying again...",flush=True)
