--- animations/draw_digits.py
@@ -1,903 +1,0 @@
-def number(number, canvas, c1, c2, c3):
-    x = 3
-    if len(str(number)) == 2:
-        n = str(number)
-        n1 = int(n[0])
-        n2 = int(n[1])
-        y_1 = 3
-        y_2 = 17
-        digit(n1, canvas, x, y_1, c1, c2, c3)
-        digit(n2, canvas, x, y_2, c1, c2, c3)
-    else:
-        y = 10
-        digit(number, canvas, x, y, c1, c2, c3)
-
-def digit(number, canvas, x, y, c1, c2, c3):
-    if number == 0:
-        zero(canvas, x, y, c1, c2, c3)
-    elif number == 1:
-        one(canvas, x, y, c1, c2, c3) 
-    elif number == 2:
-        two(canvas, x, y, c1, c2, c3) 
-    elif number == 3:
-        three(canvas, x, y, c1, c2, c3) 
-    elif number == 4:
-        four(canvas, x, y, c1, c2, c3) 
-    elif number == 5:
-        five(canvas, x, y, c1, c2, c3) 
-    elif number == 6:
-        six(canvas, x, y, c1, c2, c3) 
-    elif number == 7:
-        seven(canvas, x, y, c1, c2, c3) 
-    elif number == 8:
-        eight(canvas, x, y, c1, c2, c3) 
-    elif number == 9:
-        nine(canvas, x, y, c1, c2, c3) 
-
-def zero(canvas, x, y, c1, c2, c3):
-    canvas.SetPixel(x + 4, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 11, c1, c2, c3)
-
-def one(canvas, x, y, c1, c2, c3):
-    canvas.SetPixel(x + 0, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 9, c1, c2, c3)
-
-def two(canvas, x, y, c1, c2, c3):
-    canvas.SetPixel(x + 0, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 11, c1, c2, c3)
-
-def three(canvas, x, y, c1, c2, c3):
-    canvas.SetPixel(x + 2, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 11, c1, c2, c3)
-
-def four(canvas, x, y, c1, c2, c3):
-    canvas.SetPixel(x + 6, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 11, c1, c2, c3) 
-
-def five(canvas, x, y, c1, c2, c3):
-    canvas.SetPixel(x + 2, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 11, c1, c2, c3)
-
-def six(canvas, x, y, c1, c2, c3):
-    canvas.SetPixel(x + 2, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 11, c1, c2, c3)
-
-def seven(canvas, x, y, c1, c2, c3):
-    canvas.SetPixel(x + 18, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 11, c1, c2, c3)
-
-def eight(canvas, x, y, c1, c2, c3):
-    canvas.SetPixel(x + 2, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 11, c1, c2, c3)
-
-def nine(canvas, x, y, c1, c2, c3):
-    canvas.SetPixel(x + 12, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 0, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 1, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 2, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 3, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 4, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 5, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 11, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 11, c1, c2, c3)
