--- animations/draw_digits_transition.py
@@ -1,1252 +1,0 @@
-# Parameters to control the transition brightness
-fix1 = 80 
-fix2 = 3
-
-def number(number, canvas, c1, c2, c3, row):
-    x = -3
-    if len(str(number)) == 2:
-        n = str(number)
-        n1 = int(n[0])
-        n2 = int(n[1])
-        y_1 = 3
-        y_2 = 17
-        digit(n1, canvas, x, y_1, c1, c2, c3, row)
-        digit(n2, canvas, x, y_2, c1, c2, c3, row)
-    else:
-        y = 10
-        digit(number, canvas, x, y, c1, c2, c3, row)
-
-def digit(number, canvas, x, y, c1, c2, c3, row):
-    if number == 0:
-        zero(canvas, x, y, c1, c2, c3, row)
-    elif number == 1:
-        one(canvas, x, y, c1, c2, c3, row) 
-    elif number == 2:
-        two(canvas, x, y, c1, c2, c3, row) 
-    elif number == 3:
-        three(canvas, x, y, c1, c2, c3, row) 
-    elif number == 4:
-        four(canvas, x, y, c1, c2, c3, row) 
-    elif number == 5:
-        five(canvas, x, y, c1, c2, c3, row) 
-    elif number == 6:
-        six(canvas, x, y, c1, c2, c3, row) 
-    elif number == 7:
-        seven(canvas, x, y, c1, c2, c3, row) 
-    elif number == 8:
-        eight(canvas, x, y, c1, c2, c3, row) 
-    elif number == 9:
-        nine(canvas, x, y, c1, c2, c3, row) 
-
-def zero(canvas, x, y, o1, o2, o3, row):
-    c1,c2,c3 = min(o1,o1/(fix1/(20 + (row * fix2)))), min(o2,o2/(fix1/(20 + (row * fix2)))), min(o3,o3/(fix1/(20 + (row * fix2))))
-    canvas.SetPixel(x + 6, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(19 + (row * fix2)))), min(o2,o2/(fix1/(19 + (row * fix2)))), min(o3,o3/(fix1/(19 + (row * fix2))))
-    canvas.SetPixel(x + 7, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(18 + (row * fix2)))), min(o2,o2/(fix1/(18 + (row * fix2)))), min(o3,o3/(fix1/(18 + (row * fix2))))
-    canvas.SetPixel(x + 8, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(17 + (row * fix2)))), min(o2,o2/(fix1/(17 + (row * fix2)))), min(o3,o3/(fix1/(17 + (row * fix2))))
-    canvas.SetPixel(x + 9, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(16 + (row * fix2)))), min(o2,o2/(fix1/(16 + (row * fix2)))), min(o3,o3/(fix1/(16 + (row * fix2))))
-    canvas.SetPixel(x + 10, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(15 + (row * fix2)))), min(o2,o2/(fix1/(15 + (row * fix2)))), min(o3,o3/(fix1/(15 + (row * fix2))))
-    canvas.SetPixel(x + 11, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(14 + (row * fix2)))), min(o2,o2/(fix1/(14 + (row * fix2)))), min(o3,o3/(fix1/(14 + (row * fix2))))
-    canvas.SetPixel(x + 12, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(13 + (row * fix2)))), min(o2,o2/(fix1/(13 + (row * fix2)))), min(o3,o3/(fix1/(13 + (row * fix2))))
-    canvas.SetPixel(x + 13, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(12 + (row * fix2)))), min(o2,o2/(fix1/(12 + (row * fix2)))), min(o3,o3/(fix1/(12 + (row * fix2))))
-    canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(11 + (row * fix2)))), min(o2,o2/(fix1/(11 + (row * fix2)))), min(o3,o3/(fix1/(11 + (row * fix2))))
-    canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(10 + (row * fix2)))), min(o2,o2/(fix1/(10 + (row * fix2)))), min(o3,o3/(fix1/(10 + (row * fix2))))
-    canvas.SetPixel(x + 16, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(9 + (row * fix2)))), min(o2,o2/(fix1/(9 + (row * fix2)))), min(o3,o3/(fix1/(9 + (row * fix2))))
-    canvas.SetPixel(x + 17, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(8 + (row * fix2)))), min(o2,o2/(fix1/(8 + (row * fix2)))), min(o3,o3/(fix1/(8 + (row * fix2))))
-    canvas.SetPixel(x + 18, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(7 + (row * fix2)))), min(o2,o2/(fix1/(7 + (row * fix2)))), min(o3,o3/(fix1/(7 + (row * fix2))))
-    canvas.SetPixel(x + 19, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(6 + (row * fix2)))), min(o2,o2/(fix1/(6 + (row * fix2)))), min(o3,o3/(fix1/(6 + (row * fix2))))
-    canvas.SetPixel(x + 20, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(5 + (row * fix2)))), min(o2,o2/(fix1/(5 + (row * fix2)))), min(o3,o3/(fix1/(5 + (row * fix2))))
-    canvas.SetPixel(x + 21, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(4 + (row * fix2)))), min(o2,o2/(fix1/(4 + (row * fix2)))), min(o3,o3/(fix1/(4 + (row * fix2))))
-    canvas.SetPixel(x + 22, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(3 + (row * fix2)))), min(o2,o2/(fix1/(3 + (row * fix2)))), min(o3,o3/(fix1/(3 + (row * fix2))))
-    canvas.SetPixel(x + 23, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(2 + (row * fix2)))), min(o2,o2/(fix1/(2 + (row * fix2)))), min(o3,o3/(fix1/(2 + (row * fix2))))
-    canvas.SetPixel(x + 24, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(1 + (row * fix2)))), min(o2,o2/(fix1/(1 + (row * fix2)))), min(o3,o3/(fix1/(1 + (row * fix2))))
-    canvas.SetPixel(x + 25, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 7, c1, c2, c3)
-
-def one(canvas, x, y, o1, o2, o3, row):
-    c1,c2,c3 = min(o1,o1/(fix1/(20 + (row * fix2)))), min(o2,o2/(fix1/(20 + (row * fix2)))), min(o3,o3/(fix1/(20 + (row * fix2))))
-    canvas.SetPixel(x + 6, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(19 + (row * fix2)))), min(o2,o2/(fix1/(19 + (row * fix2)))), min(o3,o3/(fix1/(19 + (row * fix2))))
-    canvas.SetPixel(x + 7, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(18 + (row * fix2)))), min(o2,o2/(fix1/(18 + (row * fix2)))), min(o3,o3/(fix1/(18 + (row * fix2))))
-    canvas.SetPixel(x + 8, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(17 + (row * fix2)))), min(o2,o2/(fix1/(17 + (row * fix2)))), min(o3,o3/(fix1/(17 + (row * fix2))))
-    canvas.SetPixel(x + 9, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(16 + (row * fix2)))), min(o2,o2/(fix1/(16 + (row * fix2)))), min(o3,o3/(fix1/(16 + (row * fix2))))
-    canvas.SetPixel(x + 10, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(15 + (row * fix2)))), min(o2,o2/(fix1/(15 + (row * fix2)))), min(o3,o3/(fix1/(15 + (row * fix2))))
-    canvas.SetPixel(x + 11, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(14 + (row * fix2)))), min(o2,o2/(fix1/(14 + (row * fix2)))), min(o3,o3/(fix1/(14 + (row * fix2))))
-    canvas.SetPixel(x + 12, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(13 + (row * fix2)))), min(o2,o2/(fix1/(13 + (row * fix2)))), min(o3,o3/(fix1/(13 + (row * fix2))))
-    canvas.SetPixel(x + 13, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(12 + (row * fix2)))), min(o2,o2/(fix1/(12 + (row * fix2)))), min(o3,o3/(fix1/(12 + (row * fix2))))
-    canvas.SetPixel(x + 14, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(11 + (row * fix2)))), min(o2,o2/(fix1/(11 + (row * fix2)))), min(o3,o3/(fix1/(11 + (row * fix2))))
-    canvas.SetPixel(x + 15, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(10 + (row * fix2)))), min(o2,o2/(fix1/(10 + (row * fix2)))), min(o3,o3/(fix1/(10 + (row * fix2))))
-    canvas.SetPixel(x + 16, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(9 + (row * fix2)))), min(o2,o2/(fix1/(9 + (row * fix2)))), min(o3,o3/(fix1/(9 + (row * fix2))))
-    canvas.SetPixel(x + 17, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(8 + (row * fix2)))), min(o2,o2/(fix1/(8 + (row * fix2)))), min(o3,o3/(fix1/(8 + (row * fix2))))
-    canvas.SetPixel(x + 18, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(7 + (row * fix2)))), min(o2,o2/(fix1/(7 + (row * fix2)))), min(o3,o3/(fix1/(7 + (row * fix2))))
-    canvas.SetPixel(x + 19, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(6 + (row * fix2)))), min(o2,o2/(fix1/(6 + (row * fix2)))), min(o3,o3/(fix1/(6 + (row * fix2))))
-    canvas.SetPixel(x + 20, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(5 + (row * fix2)))), min(o2,o2/(fix1/(5 + (row * fix2)))), min(o3,o3/(fix1/(5 + (row * fix2))))
-    canvas.SetPixel(x + 21, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(4 + (row * fix2)))), min(o2,o2/(fix1/(4 + (row * fix2)))), min(o3,o3/(fix1/(4 + (row * fix2))))
-    canvas.SetPixel(x + 22, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(3 + (row * fix2)))), min(o2,o2/(fix1/(3 + (row * fix2)))), min(o3,o3/(fix1/(3 + (row * fix2))))
-    canvas.SetPixel(x + 23, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(2 + (row * fix2)))), min(o2,o2/(fix1/(2 + (row * fix2)))), min(o3,o3/(fix1/(2 + (row * fix2))))
-    canvas.SetPixel(x + 24, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(1 + (row * fix2)))), min(o2,o2/(fix1/(1 + (row * fix2)))), min(o3,o3/(fix1/(1 + (row * fix2))))
-    canvas.SetPixel(x + 25, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 5, c1, c2, c3)
-
-def two(canvas, x, y, o1, o2, o3, row):
-    c1,c2,c3 = min(o1,o1/(fix1/(20 + (row * fix2)))), min(o2,o2/(fix1/(20 + (row * fix2)))), min(o3,o3/(fix1/(20 + (row * fix2))))
-    canvas.SetPixel(x + 6, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(19 + (row * fix2)))), min(o2,o2/(fix1/(19 + (row * fix2)))), min(o3,o3/(fix1/(19 + (row * fix2))))
-    canvas.SetPixel(x + 7, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(18 + (row * fix2)))), min(o2,o2/(fix1/(18 + (row * fix2)))), min(o3,o3/(fix1/(18 + (row * fix2))))
-    canvas.SetPixel(x + 8, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(17 + (row * fix2)))), min(o2,o2/(fix1/(17 + (row * fix2)))), min(o3,o3/(fix1/(17 + (row * fix2))))
-    canvas.SetPixel(x + 9, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(16 + (row * fix2)))), min(o2,o2/(fix1/(16 + (row * fix2)))), min(o3,o3/(fix1/(16 + (row * fix2))))
-    canvas.SetPixel(x + 10, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(15 + (row * fix2)))), min(o2,o2/(fix1/(15 + (row * fix2)))), min(o3,o3/(fix1/(15 + (row * fix2))))
-    canvas.SetPixel(x + 11, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(14 + (row * fix2)))), min(o2,o2/(fix1/(14 + (row * fix2)))), min(o3,o3/(fix1/(14 + (row * fix2))))
-    canvas.SetPixel(x + 12, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 3, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(13 + (row * fix2)))), min(o2,o2/(fix1/(13 + (row * fix2)))), min(o3,o3/(fix1/(13 + (row * fix2))))
-    canvas.SetPixel(x + 13, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 3, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(12 + (row * fix2)))), min(o2,o2/(fix1/(12 + (row * fix2)))), min(o3,o3/(fix1/(12 + (row * fix2))))
-    canvas.SetPixel(x + 14, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(11 + (row * fix2)))), min(o2,o2/(fix1/(11 + (row * fix2)))), min(o3,o3/(fix1/(11 + (row * fix2))))
-    canvas.SetPixel(x + 15, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 5, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(10 + (row * fix2)))), min(o2,o2/(fix1/(10 + (row * fix2)))), min(o3,o3/(fix1/(10 + (row * fix2))))
-    canvas.SetPixel(x + 16, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(9 + (row * fix2)))), min(o2,o2/(fix1/(9 + (row * fix2)))), min(o3,o3/(fix1/(9 + (row * fix2))))
-    canvas.SetPixel(x + 17, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(8 + (row * fix2)))), min(o2,o2/(fix1/(8 + (row * fix2)))), min(o3,o3/(fix1/(8 + (row * fix2))))
-    canvas.SetPixel(x + 18, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(7 + (row * fix2)))), min(o2,o2/(fix1/(7 + (row * fix2)))), min(o3,o3/(fix1/(7 + (row * fix2))))
-    canvas.SetPixel(x + 19, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(6 + (row * fix2)))), min(o2,o2/(fix1/(6 + (row * fix2)))), min(o3,o3/(fix1/(6 + (row * fix2))))
-    canvas.SetPixel(x + 20, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(5 + (row * fix2)))), min(o2,o2/(fix1/(5 + (row * fix2)))), min(o3,o3/(fix1/(5 + (row * fix2))))
-    canvas.SetPixel(x + 21, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(4 + (row * fix2)))), min(o2,o2/(fix1/(4 + (row * fix2)))), min(o3,o3/(fix1/(4 + (row * fix2))))
-    canvas.SetPixel(x + 22, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(3 + (row * fix2)))), min(o2,o2/(fix1/(3 + (row * fix2)))), min(o3,o3/(fix1/(3 + (row * fix2))))
-    canvas.SetPixel(x + 23, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(2 + (row * fix2)))), min(o2,o2/(fix1/(2 + (row * fix2)))), min(o3,o3/(fix1/(2 + (row * fix2))))
-    canvas.SetPixel(x + 24, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(1 + (row * fix2)))), min(o2,o2/(fix1/(1 + (row * fix2)))), min(o3,o3/(fix1/(1 + (row * fix2))))
-    canvas.SetPixel(x + 25, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 9, c1, c2, c3)
-
-def three(canvas, x, y, o1, o2, o3, row):
-    c1,c2,c3 = min(o1,o1/(fix1/(20 + (row * fix2)))), min(o2,o2/(fix1/(20 + (row * fix2)))), min(o3,o3/(fix1/(20 + (row * fix2))))
-    canvas.SetPixel(x + 6, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(19 + (row * fix2)))), min(o2,o2/(fix1/(19 + (row * fix2)))), min(o3,o3/(fix1/(19 + (row * fix2))))
-    canvas.SetPixel(x + 7, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(18 + (row * fix2)))), min(o2,o2/(fix1/(18 + (row * fix2)))), min(o3,o3/(fix1/(18 + (row * fix2))))
-    canvas.SetPixel(x + 8, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(17 + (row * fix2)))), min(o2,o2/(fix1/(17 + (row * fix2)))), min(o3,o3/(fix1/(17 + (row * fix2))))
-    canvas.SetPixel(x + 9, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(16 + (row * fix2)))), min(o2,o2/(fix1/(16 + (row * fix2)))), min(o3,o3/(fix1/(16 + (row * fix2))))
-    canvas.SetPixel(x + 10, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(15 + (row * fix2)))), min(o2,o2/(fix1/(15 + (row * fix2)))), min(o3,o3/(fix1/(15 + (row * fix2))))
-    canvas.SetPixel(x + 11, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(14 + (row * fix2)))), min(o2,o2/(fix1/(14 + (row * fix2)))), min(o3,o3/(fix1/(14 + (row * fix2))))
-    canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(13 + (row * fix2)))), min(o2,o2/(fix1/(13 + (row * fix2)))), min(o3,o3/(fix1/(13 + (row * fix2))))
-    canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(12 + (row * fix2)))), min(o2,o2/(fix1/(12 + (row * fix2)))), min(o3,o3/(fix1/(12 + (row * fix2))))
-    canvas.SetPixel(x + 14, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(11 + (row * fix2)))), min(o2,o2/(fix1/(11 + (row * fix2)))), min(o3,o3/(fix1/(11 + (row * fix2))))
-    canvas.SetPixel(x + 15, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(10 + (row * fix2)))), min(o2,o2/(fix1/(10 + (row * fix2)))), min(o3,o3/(fix1/(10 + (row * fix2))))
-    canvas.SetPixel(x + 16, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(9 + (row * fix2)))), min(o2,o2/(fix1/(9 + (row * fix2)))), min(o3,o3/(fix1/(9 + (row * fix2))))
-    canvas.SetPixel(x + 17, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(8 + (row * fix2)))), min(o2,o2/(fix1/(8 + (row * fix2)))), min(o3,o3/(fix1/(8 + (row * fix2))))
-    canvas.SetPixel(x + 18, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(7 + (row * fix2)))), min(o2,o2/(fix1/(7 + (row * fix2)))), min(o3,o3/(fix1/(7 + (row * fix2))))
-    canvas.SetPixel(x + 19, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(6 + (row * fix2)))), min(o2,o2/(fix1/(6 + (row * fix2)))), min(o3,o3/(fix1/(6 + (row * fix2))))
-    canvas.SetPixel(x + 20, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(5 + (row * fix2)))), min(o2,o2/(fix1/(5 + (row * fix2)))), min(o3,o3/(fix1/(5 + (row * fix2))))
-    canvas.SetPixel(x + 21, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(4 + (row * fix2)))), min(o2,o2/(fix1/(4 + (row * fix2)))), min(o3,o3/(fix1/(4 + (row * fix2))))
-    canvas.SetPixel(x + 22, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(3 + (row * fix2)))), min(o2,o2/(fix1/(3 + (row * fix2)))), min(o3,o3/(fix1/(3 + (row * fix2))))
-    canvas.SetPixel(x + 23, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(2 + (row * fix2)))), min(o2,o2/(fix1/(2 + (row * fix2)))), min(o3,o3/(fix1/(2 + (row * fix2))))
-    canvas.SetPixel(x + 24, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(1 + (row * fix2)))), min(o2,o2/(fix1/(1 + (row * fix2)))), min(o3,o3/(fix1/(1 + (row * fix2))))
-    canvas.SetPixel(x + 25, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 9, c1, c2, c3)
-
-def four(canvas, x, y, o1, o2, o3, row):
-    c1,c2,c3 = min(o1,o1/(fix1/(20 + (row * fix2)))), min(o2,o2/(fix1/(20 + (row * fix2)))), min(o3,o3/(fix1/(20 + (row * fix2))))
-    canvas.SetPixel(x + 6, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(19 + (row * fix2)))), min(o2,o2/(fix1/(19 + (row * fix2)))), min(o3,o3/(fix1/(19 + (row * fix2))))
-    canvas.SetPixel(x + 7, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(18 + (row * fix2)))), min(o2,o2/(fix1/(18 + (row * fix2)))), min(o3,o3/(fix1/(18 + (row * fix2))))
-    canvas.SetPixel(x + 8, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(17 + (row * fix2)))), min(o2,o2/(fix1/(17 + (row * fix2)))), min(o3,o3/(fix1/(17 + (row * fix2))))
-    canvas.SetPixel(x + 9, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(16 + (row * fix2)))), min(o2,o2/(fix1/(16 + (row * fix2)))), min(o3,o3/(fix1/(16 + (row * fix2))))
-    canvas.SetPixel(x + 10, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(15 + (row * fix2)))), min(o2,o2/(fix1/(15 + (row * fix2)))), min(o3,o3/(fix1/(15 + (row * fix2))))
-    canvas.SetPixel(x + 11, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(14 + (row * fix2)))), min(o2,o2/(fix1/(14 + (row * fix2)))), min(o3,o3/(fix1/(14 + (row * fix2))))
-    canvas.SetPixel(x + 12, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(13 + (row * fix2)))), min(o2,o2/(fix1/(13 + (row * fix2)))), min(o3,o3/(fix1/(13 + (row * fix2))))
-    canvas.SetPixel(x + 13, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(12 + (row * fix2)))), min(o2,o2/(fix1/(12 + (row * fix2)))), min(o3,o3/(fix1/(12 + (row * fix2))))
-    canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(11 + (row * fix2)))), min(o2,o2/(fix1/(11 + (row * fix2)))), min(o3,o3/(fix1/(11 + (row * fix2))))
-    canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(10 + (row * fix2)))), min(o2,o2/(fix1/(10 + (row * fix2)))), min(o3,o3/(fix1/(10 + (row * fix2))))
-    canvas.SetPixel(x + 16, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(9 + (row * fix2)))), min(o2,o2/(fix1/(9 + (row * fix2)))), min(o3,o3/(fix1/(9 + (row * fix2))))
-    canvas.SetPixel(x + 17, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(8 + (row * fix2)))), min(o2,o2/(fix1/(8 + (row * fix2)))), min(o3,o3/(fix1/(8 + (row * fix2))))
-    canvas.SetPixel(x + 18, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(7 + (row * fix2)))), min(o2,o2/(fix1/(7 + (row * fix2)))), min(o3,o3/(fix1/(7 + (row * fix2))))
-    canvas.SetPixel(x + 19, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(6 + (row * fix2)))), min(o2,o2/(fix1/(6 + (row * fix2)))), min(o3,o3/(fix1/(6 + (row * fix2))))
-    canvas.SetPixel(x + 20, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(5 + (row * fix2)))), min(o2,o2/(fix1/(5 + (row * fix2)))), min(o3,o3/(fix1/(5 + (row * fix2))))
-    canvas.SetPixel(x + 21, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(4 + (row * fix2)))), min(o2,o2/(fix1/(4 + (row * fix2)))), min(o3,o3/(fix1/(4 + (row * fix2))))
-    canvas.SetPixel(x + 22, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(3 + (row * fix2)))), min(o2,o2/(fix1/(3 + (row * fix2)))), min(o3,o3/(fix1/(3 + (row * fix2))))
-    canvas.SetPixel(x + 23, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(2 + (row * fix2)))), min(o2,o2/(fix1/(2 + (row * fix2)))), min(o3,o3/(fix1/(2 + (row * fix2))))
-    canvas.SetPixel(x + 24, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(1 + (row * fix2)))), min(o2,o2/(fix1/(1 + (row * fix2)))), min(o3,o3/(fix1/(1 + (row * fix2))))
-    canvas.SetPixel(x + 25, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 9, c1, c2, c3)
-
-def five(canvas, x, y, o1, o2, o3, row):
-    c1,c2,c3 = min(o1,o1/(fix1/(20 + (row * fix2)))), min(o2,o2/(fix1/(20 + (row * fix2)))), min(o3,o3/(fix1/(20 + (row * fix2))))
-    canvas.SetPixel(x + 6, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(19 + (row * fix2)))), min(o2,o2/(fix1/(19 + (row * fix2)))), min(o3,o3/(fix1/(19 + (row * fix2))))
-    canvas.SetPixel(x + 7, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(18 + (row * fix2)))), min(o2,o2/(fix1/(18 + (row * fix2)))), min(o3,o3/(fix1/(18 + (row * fix2))))
-    canvas.SetPixel(x + 8, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(17 + (row * fix2)))), min(o2,o2/(fix1/(17 + (row * fix2)))), min(o3,o3/(fix1/(17 + (row * fix2))))
-    canvas.SetPixel(x + 9, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(16 + (row * fix2)))), min(o2,o2/(fix1/(16 + (row * fix2)))), min(o3,o3/(fix1/(16 + (row * fix2))))
-    canvas.SetPixel(x + 10, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(15 + (row * fix2)))), min(o2,o2/(fix1/(15 + (row * fix2)))), min(o3,o3/(fix1/(15 + (row * fix2))))
-    canvas.SetPixel(x + 11, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(14 + (row * fix2)))), min(o2,o2/(fix1/(14 + (row * fix2)))), min(o3,o3/(fix1/(14 + (row * fix2))))
-    canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(13 + (row * fix2)))), min(o2,o2/(fix1/(13 + (row * fix2)))), min(o3,o3/(fix1/(13 + (row * fix2))))
-    canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(12 + (row * fix2)))), min(o2,o2/(fix1/(12 + (row * fix2)))), min(o3,o3/(fix1/(12 + (row * fix2))))
-    canvas.SetPixel(x + 14, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(11 + (row * fix2)))), min(o2,o2/(fix1/(11 + (row * fix2)))), min(o3,o3/(fix1/(11 + (row * fix2))))
-    canvas.SetPixel(x + 15, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(10 + (row * fix2)))), min(o2,o2/(fix1/(10 + (row * fix2)))), min(o3,o3/(fix1/(10 + (row * fix2))))
-    canvas.SetPixel(x + 16, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(9 + (row * fix2)))), min(o2,o2/(fix1/(9 + (row * fix2)))), min(o3,o3/(fix1/(9 + (row * fix2))))
-    canvas.SetPixel(x + 17, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(8 + (row * fix2)))), min(o2,o2/(fix1/(8 + (row * fix2)))), min(o3,o3/(fix1/(8 + (row * fix2))))
-    canvas.SetPixel(x + 18, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(7 + (row * fix2)))), min(o2,o2/(fix1/(7 + (row * fix2)))), min(o3,o3/(fix1/(7 + (row * fix2))))
-    canvas.SetPixel(x + 19, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(6 + (row * fix2)))), min(o2,o2/(fix1/(6 + (row * fix2)))), min(o3,o3/(fix1/(6 + (row * fix2))))
-    canvas.SetPixel(x + 20, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(5 + (row * fix2)))), min(o2,o2/(fix1/(5 + (row * fix2)))), min(o3,o3/(fix1/(5 + (row * fix2))))
-    canvas.SetPixel(x + 21, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(4 + (row * fix2)))), min(o2,o2/(fix1/(4 + (row * fix2)))), min(o3,o3/(fix1/(4 + (row * fix2))))
-    canvas.SetPixel(x + 22, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(3 + (row * fix2)))), min(o2,o2/(fix1/(3 + (row * fix2)))), min(o3,o3/(fix1/(3 + (row * fix2))))
-    canvas.SetPixel(x + 23, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(2 + (row * fix2)))), min(o2,o2/(fix1/(2 + (row * fix2)))), min(o3,o3/(fix1/(2 + (row * fix2))))
-    canvas.SetPixel(x + 24, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(1 + (row * fix2)))), min(o2,o2/(fix1/(1 + (row * fix2)))), min(o3,o3/(fix1/(1 + (row * fix2))))
-    canvas.SetPixel(x + 25, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 11, c1, c2, c3)
-
-def six(canvas, x, y, o1, o2, o3, row):
-    c1,c2,c3 = min(o1,o1/(fix1/(20 + (row * fix2)))), min(o2,o2/(fix1/(20 + (row * fix2)))), min(o3,o3/(fix1/(20 + (row * fix2))))
-    canvas.SetPixel(x + 6, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(19 + (row * fix2)))), min(o2,o2/(fix1/(19 + (row * fix2)))), min(o3,o3/(fix1/(19 + (row * fix2))))
-    canvas.SetPixel(x + 7, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(18 + (row * fix2)))), min(o2,o2/(fix1/(18 + (row * fix2)))), min(o3,o3/(fix1/(18 + (row * fix2))))
-    canvas.SetPixel(x + 8, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(17 + (row * fix2)))), min(o2,o2/(fix1/(17 + (row * fix2)))), min(o3,o3/(fix1/(17 + (row * fix2))))
-    canvas.SetPixel(x + 9, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(16 + (row * fix2)))), min(o2,o2/(fix1/(16 + (row * fix2)))), min(o3,o3/(fix1/(16 + (row * fix2))))
-    canvas.SetPixel(x + 10, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(15 + (row * fix2)))), min(o2,o2/(fix1/(15 + (row * fix2)))), min(o3,o3/(fix1/(15 + (row * fix2))))
-    canvas.SetPixel(x + 11, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(14 + (row * fix2)))), min(o2,o2/(fix1/(14 + (row * fix2)))), min(o3,o3/(fix1/(14 + (row * fix2))))
-    canvas.SetPixel(x + 12, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(13 + (row * fix2)))), min(o2,o2/(fix1/(13 + (row * fix2)))), min(o3,o3/(fix1/(13 + (row * fix2))))
-    canvas.SetPixel(x + 13, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(12 + (row * fix2)))), min(o2,o2/(fix1/(12 + (row * fix2)))), min(o3,o3/(fix1/(12 + (row * fix2))))
-    canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(11 + (row * fix2)))), min(o2,o2/(fix1/(11 + (row * fix2)))), min(o3,o3/(fix1/(11 + (row * fix2))))
-    canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(10 + (row * fix2)))), min(o2,o2/(fix1/(10 + (row * fix2)))), min(o3,o3/(fix1/(10 + (row * fix2))))
-    canvas.SetPixel(x + 16, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(9 + (row * fix2)))), min(o2,o2/(fix1/(9 + (row * fix2)))), min(o3,o3/(fix1/(9 + (row * fix2))))
-    canvas.SetPixel(x + 17, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(8 + (row * fix2)))), min(o2,o2/(fix1/(8 + (row * fix2)))), min(o3,o3/(fix1/(8 + (row * fix2))))
-    canvas.SetPixel(x + 18, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(7 + (row * fix2)))), min(o2,o2/(fix1/(7 + (row * fix2)))), min(o3,o3/(fix1/(7 + (row * fix2))))
-    canvas.SetPixel(x + 19, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(6 + (row * fix2)))), min(o2,o2/(fix1/(6 + (row * fix2)))), min(o3,o3/(fix1/(6 + (row * fix2))))
-    canvas.SetPixel(x + 20, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(5 + (row * fix2)))), min(o2,o2/(fix1/(5 + (row * fix2)))), min(o3,o3/(fix1/(5 + (row * fix2))))
-    canvas.SetPixel(x + 21, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 1, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(4 + (row * fix2)))), min(o2,o2/(fix1/(4 + (row * fix2)))), min(o3,o3/(fix1/(4 + (row * fix2))))
-    canvas.SetPixel(x + 22, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 3, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(3 + (row * fix2)))), min(o2,o2/(fix1/(3 + (row * fix2)))), min(o3,o3/(fix1/(3 + (row * fix2))))
-    canvas.SetPixel(x + 23, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 3, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(2 + (row * fix2)))), min(o2,o2/(fix1/(2 + (row * fix2)))), min(o3,o3/(fix1/(2 + (row * fix2))))
-    canvas.SetPixel(x + 24, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(1 + (row * fix2)))), min(o2,o2/(fix1/(1 + (row * fix2)))), min(o3,o3/(fix1/(1 + (row * fix2))))
-    canvas.SetPixel(x + 25, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 9, c1, c2, c3)
-
-def seven(canvas, x, y, o1, o2, o3, row):
-    c1,c2,c3 = min(o1,o1/(fix1/(20 + (row * fix2)))), min(o2,o2/(fix1/(20 + (row * fix2)))), min(o3,o3/(fix1/(20 + (row * fix2))))
-    canvas.SetPixel(x + 6, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(19 + (row * fix2)))), min(o2,o2/(fix1/(19 + (row * fix2)))), min(o3,o3/(fix1/(19 + (row * fix2))))
-    canvas.SetPixel(x + 7, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(18 + (row * fix2)))), min(o2,o2/(fix1/(18 + (row * fix2)))), min(o3,o3/(fix1/(18 + (row * fix2))))
-    canvas.SetPixel(x + 8, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(17 + (row * fix2)))), min(o2,o2/(fix1/(17 + (row * fix2)))), min(o3,o3/(fix1/(17 + (row * fix2))))
-    canvas.SetPixel(x + 9, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(16 + (row * fix2)))), min(o2,o2/(fix1/(16 + (row * fix2)))), min(o3,o3/(fix1/(16 + (row * fix2))))
-    canvas.SetPixel(x + 10, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(15 + (row * fix2)))), min(o2,o2/(fix1/(15 + (row * fix2)))), min(o3,o3/(fix1/(15 + (row * fix2))))
-    canvas.SetPixel(x + 11, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(14 + (row * fix2)))), min(o2,o2/(fix1/(14 + (row * fix2)))), min(o3,o3/(fix1/(14 + (row * fix2))))
-    canvas.SetPixel(x + 12, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(13 + (row * fix2)))), min(o2,o2/(fix1/(13 + (row * fix2)))), min(o3,o3/(fix1/(13 + (row * fix2))))
-    canvas.SetPixel(x + 13, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(12 + (row * fix2)))), min(o2,o2/(fix1/(12 + (row * fix2)))), min(o3,o3/(fix1/(12 + (row * fix2))))
-    canvas.SetPixel(x + 14, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(11 + (row * fix2)))), min(o2,o2/(fix1/(11 + (row * fix2)))), min(o3,o3/(fix1/(11 + (row * fix2))))
-    canvas.SetPixel(x + 15, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(10 + (row * fix2)))), min(o2,o2/(fix1/(10 + (row * fix2)))), min(o3,o3/(fix1/(10 + (row * fix2))))
-    canvas.SetPixel(x + 16, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(9 + (row * fix2)))), min(o2,o2/(fix1/(9 + (row * fix2)))), min(o3,o3/(fix1/(9 + (row * fix2))))
-    canvas.SetPixel(x + 17, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(8 + (row * fix2)))), min(o2,o2/(fix1/(8 + (row * fix2)))), min(o3,o3/(fix1/(8 + (row * fix2))))
-    canvas.SetPixel(x + 18, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(7 + (row * fix2)))), min(o2,o2/(fix1/(7 + (row * fix2)))), min(o3,o3/(fix1/(7 + (row * fix2))))
-    canvas.SetPixel(x + 19, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(6 + (row * fix2)))), min(o2,o2/(fix1/(6 + (row * fix2)))), min(o3,o3/(fix1/(6 + (row * fix2))))
-    canvas.SetPixel(x + 20, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(5 + (row * fix2)))), min(o2,o2/(fix1/(5 + (row * fix2)))), min(o3,o3/(fix1/(5 + (row * fix2))))
-    canvas.SetPixel(x + 21, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(4 + (row * fix2)))), min(o2,o2/(fix1/(4 + (row * fix2)))), min(o3,o3/(fix1/(4 + (row * fix2))))
-    canvas.SetPixel(x + 22, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(3 + (row * fix2)))), min(o2,o2/(fix1/(3 + (row * fix2)))), min(o3,o3/(fix1/(3 + (row * fix2))))
-    canvas.SetPixel(x + 23, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(2 + (row * fix2)))), min(o2,o2/(fix1/(2 + (row * fix2)))), min(o3,o3/(fix1/(2 + (row * fix2))))
-    canvas.SetPixel(x + 24, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(1 + (row * fix2)))), min(o2,o2/(fix1/(1 + (row * fix2)))), min(o3,o3/(fix1/(1 + (row * fix2))))
-    canvas.SetPixel(x + 25, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 11, c1, c2, c3)
-
-def eight(canvas, x, y, o1, o2, o3, row):
-    c1,c2,c3 = min(o1,o1/(fix1/(20 + (row * fix2)))), min(o2,o2/(fix1/(20 + (row * fix2)))), min(o3,o3/(fix1/(20 + (row * fix2))))
-    canvas.SetPixel(x + 6, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(19 + (row * fix2)))), min(o2,o2/(fix1/(19 + (row * fix2)))), min(o3,o3/(fix1/(19 + (row * fix2))))
-    canvas.SetPixel(x + 7, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(18 + (row * fix2)))), min(o2,o2/(fix1/(18 + (row * fix2)))), min(o3,o3/(fix1/(18 + (row * fix2))))
-    canvas.SetPixel(x + 8, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(17 + (row * fix2)))), min(o2,o2/(fix1/(17 + (row * fix2)))), min(o3,o3/(fix1/(17 + (row * fix2))))
-    canvas.SetPixel(x + 9, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(16 + (row * fix2)))), min(o2,o2/(fix1/(16 + (row * fix2)))), min(o3,o3/(fix1/(16 + (row * fix2))))
-    canvas.SetPixel(x + 10, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(15 + (row * fix2)))), min(o2,o2/(fix1/(15 + (row * fix2)))), min(o3,o3/(fix1/(15 + (row * fix2))))
-    canvas.SetPixel(x + 11, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(14 + (row * fix2)))), min(o2,o2/(fix1/(14 + (row * fix2)))), min(o3,o3/(fix1/(14 + (row * fix2))))
-    canvas.SetPixel(x + 12, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(13 + (row * fix2)))), min(o2,o2/(fix1/(13 + (row * fix2)))), min(o3,o3/(fix1/(13 + (row * fix2))))
-    canvas.SetPixel(x + 13, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(12 + (row * fix2)))), min(o2,o2/(fix1/(12 + (row * fix2)))), min(o3,o3/(fix1/(12 + (row * fix2))))
-    canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(11 + (row * fix2)))), min(o2,o2/(fix1/(11 + (row * fix2)))), min(o3,o3/(fix1/(11 + (row * fix2))))
-    canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(10 + (row * fix2)))), min(o2,o2/(fix1/(10 + (row * fix2)))), min(o3,o3/(fix1/(10 + (row * fix2))))
-    canvas.SetPixel(x + 16, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(9 + (row * fix2)))), min(o2,o2/(fix1/(9 + (row * fix2)))), min(o3,o3/(fix1/(9 + (row * fix2))))
-    canvas.SetPixel(x + 17, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(8 + (row * fix2)))), min(o2,o2/(fix1/(8 + (row * fix2)))), min(o3,o3/(fix1/(8 + (row * fix2))))
-    canvas.SetPixel(x + 18, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(7 + (row * fix2)))), min(o2,o2/(fix1/(7 + (row * fix2)))), min(o3,o3/(fix1/(7 + (row * fix2))))
-    canvas.SetPixel(x + 19, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(6 + (row * fix2)))), min(o2,o2/(fix1/(6 + (row * fix2)))), min(o3,o3/(fix1/(6 + (row * fix2))))
-    canvas.SetPixel(x + 20, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(5 + (row * fix2)))), min(o2,o2/(fix1/(5 + (row * fix2)))), min(o3,o3/(fix1/(5 + (row * fix2))))
-    canvas.SetPixel(x + 21, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(4 + (row * fix2)))), min(o2,o2/(fix1/(4 + (row * fix2)))), min(o3,o3/(fix1/(4 + (row * fix2))))
-    canvas.SetPixel(x + 22, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(3 + (row * fix2)))), min(o2,o2/(fix1/(3 + (row * fix2)))), min(o3,o3/(fix1/(3 + (row * fix2))))
-    canvas.SetPixel(x + 23, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(2 + (row * fix2)))), min(o2,o2/(fix1/(2 + (row * fix2)))), min(o3,o3/(fix1/(2 + (row * fix2))))
-    canvas.SetPixel(x + 24, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(1 + (row * fix2)))), min(o2,o2/(fix1/(1 + (row * fix2)))), min(o3,o3/(fix1/(1 + (row * fix2))))
-    canvas.SetPixel(x + 25, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 9, c1, c2, c3)
-
-def nine(canvas, x, y, o1, o2, o3, row):
-    c1,c2,c3 = min(o1,o1/(fix1/(20 + (row * fix2)))), min(o2,o2/(fix1/(20 + (row * fix2)))), min(o3,o3/(fix1/(20 + (row * fix2))))
-    canvas.SetPixel(x + 6, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 6, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(19 + (row * fix2)))), min(o2,o2/(fix1/(19 + (row * fix2)))), min(o3,o3/(fix1/(19 + (row * fix2))))
-    canvas.SetPixel(x + 7, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 7, y + 7, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(18 + (row * fix2)))), min(o2,o2/(fix1/(18 + (row * fix2)))), min(o3,o3/(fix1/(18 + (row * fix2))))
-    canvas.SetPixel(x + 8, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 8, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(17 + (row * fix2)))), min(o2,o2/(fix1/(17 + (row * fix2)))), min(o3,o3/(fix1/(17 + (row * fix2))))
-    canvas.SetPixel(x + 9, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 9, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(16 + (row * fix2)))), min(o2,o2/(fix1/(16 + (row * fix2)))), min(o3,o3/(fix1/(16 + (row * fix2))))
-    canvas.SetPixel(x + 10, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 10, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(15 + (row * fix2)))), min(o2,o2/(fix1/(15 + (row * fix2)))), min(o3,o3/(fix1/(15 + (row * fix2))))
-    canvas.SetPixel(x + 11, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 11, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(14 + (row * fix2)))), min(o2,o2/(fix1/(14 + (row * fix2)))), min(o3,o3/(fix1/(14 + (row * fix2))))
-    canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(13 + (row * fix2)))), min(o2,o2/(fix1/(13 + (row * fix2)))), min(o3,o3/(fix1/(13 + (row * fix2))))
-    canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(12 + (row * fix2)))), min(o2,o2/(fix1/(12 + (row * fix2)))), min(o3,o3/(fix1/(12 + (row * fix2))))
-    canvas.SetPixel(x + 14, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 14, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(11 + (row * fix2)))), min(o2,o2/(fix1/(11 + (row * fix2)))), min(o3,o3/(fix1/(11 + (row * fix2))))
-    canvas.SetPixel(x + 15, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 15, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(10 + (row * fix2)))), min(o2,o2/(fix1/(10 + (row * fix2)))), min(o3,o3/(fix1/(10 + (row * fix2))))
-    canvas.SetPixel(x + 16, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 16, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(9 + (row * fix2)))), min(o2,o2/(fix1/(9 + (row * fix2)))), min(o3,o3/(fix1/(9 + (row * fix2))))
-    canvas.SetPixel(x + 17, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 9, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 17, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(8 + (row * fix2)))), min(o2,o2/(fix1/(8 + (row * fix2)))), min(o3,o3/(fix1/(8 + (row * fix2))))
-    canvas.SetPixel(x + 18, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 18, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(7 + (row * fix2)))), min(o2,o2/(fix1/(7 + (row * fix2)))), min(o3,o3/(fix1/(7 + (row * fix2))))
-    canvas.SetPixel(x + 19, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 19, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(6 + (row * fix2)))), min(o2,o2/(fix1/(6 + (row * fix2)))), min(o3,o3/(fix1/(6 + (row * fix2))))
-    canvas.SetPixel(x + 20, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 20, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(5 + (row * fix2)))), min(o2,o2/(fix1/(5 + (row * fix2)))), min(o3,o3/(fix1/(5 + (row * fix2))))
-    canvas.SetPixel(x + 21, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 21, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(4 + (row * fix2)))), min(o2,o2/(fix1/(4 + (row * fix2)))), min(o3,o3/(fix1/(4 + (row * fix2))))
-    canvas.SetPixel(x + 22, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 22, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(3 + (row * fix2)))), min(o2,o2/(fix1/(3 + (row * fix2)))), min(o3,o3/(fix1/(3 + (row * fix2))))
-    canvas.SetPixel(x + 23, y + 0, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 1, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 10, c1, c2, c3)
-    canvas.SetPixel(x + 23, y + 11, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(2 + (row * fix2)))), min(o2,o2/(fix1/(2 + (row * fix2)))), min(o3,o3/(fix1/(2 + (row * fix2))))
-    canvas.SetPixel(x + 24, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 24, y + 9, c1, c2, c3)
-    c1,c2,c3 = min(o1,o1/(fix1/(1 + (row * fix2)))), min(o2,o2/(fix1/(1 + (row * fix2)))), min(o3,o3/(fix1/(1 + (row * fix2))))
-    canvas.SetPixel(x + 25, y + 2, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 3, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 4, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 5, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 6, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 7, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 8, c1, c2, c3)
-    canvas.SetPixel(x + 25, y + 9, c1, c2, c3)
-
-# def four(canvas, x, y, o1, o2, o3, row):
-#     #fix1 = 80 # Higher -> Darker
-#     #fix2 = 2
-
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(20 + (row * fix2)))), min(o2,o2/(fix1/(20 + (row * fix2)))), min(o3,o3/(fix1/(20 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 6, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 6, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(19 + (row * fix2)))), min(o2,o2/(fix1/(19 + (row * fix2)))), min(o3,o3/(fix1/(19 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 7, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 7, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(18 + (row * fix2)))), min(o2,o2/(fix1/(18 + (row * fix2)))), min(o3,o3/(fix1/(18 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 8, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 8, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(17 + (row * fix2)))), min(o2,o2/(fix1/(17 + (row * fix2)))), min(o3,o3/(fix1/(17 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 9, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 9, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(16 + (row * fix2)))), min(o2,o2/(fix1/(16 + (row * fix2)))), min(o3,o3/(fix1/(16 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 10, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 10, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(15 + (row * fix2)))), min(o2,o2/(fix1/(15 + (row * fix2)))), min(o3,o3/(fix1/(15 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 11, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 11, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(14 + (row * fix2)))), min(o2,o2/(fix1/(14 + (row * fix2)))), min(o3,o3/(fix1/(14 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 12, y + 0, c1, c2, c3)
-#     canvas.SetPixel(x + 12, y + 1, c1, c2, c3)
-#     canvas.SetPixel(x + 12, y + 2, c1, c2, c3)
-#     canvas.SetPixel(x + 12, y + 3, c1, c2, c3)
-#     canvas.SetPixel(x + 12, y + 4, c1, c2, c3)
-#     canvas.SetPixel(x + 12, y + 5, c1, c2, c3)
-#     canvas.SetPixel(x + 12, y + 6, c1, c2, c3)
-#     canvas.SetPixel(x + 12, y + 7, c1, c2, c3)
-#     canvas.SetPixel(x + 12, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 12, y + 9, c1, c2, c3)
-#     canvas.SetPixel(x + 12, y + 10, c1, c2, c3)
-#     canvas.SetPixel(x + 12, y + 11, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(13 + (row * fix2)))), min(o2,o2/(fix1/(13 + (row * fix2)))), min(o3,o3/(fix1/(13 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 13, y + 0, c1, c2, c3)
-#     canvas.SetPixel(x + 13, y + 1, c1, c2, c3)
-#     canvas.SetPixel(x + 13, y + 2, c1, c2, c3)
-#     canvas.SetPixel(x + 13, y + 3, c1, c2, c3)
-#     canvas.SetPixel(x + 13, y + 4, c1, c2, c3)
-#     canvas.SetPixel(x + 13, y + 5, c1, c2, c3)
-#     canvas.SetPixel(x + 13, y + 6, c1, c2, c3)
-#     canvas.SetPixel(x + 13, y + 7, c1, c2, c3)
-#     canvas.SetPixel(x + 13, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 13, y + 9, c1, c2, c3)
-#     canvas.SetPixel(x + 13, y + 10, c1, c2, c3)
-#     canvas.SetPixel(x + 13, y + 11, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(12 + (row * fix2)))), min(o2,o2/(fix1/(12 + (row * fix2)))), min(o3,o3/(fix1/(12 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 14, y + 0, c1, c2, c3)
-#     canvas.SetPixel(x + 14, y + 1, c1, c2, c3)
-#     canvas.SetPixel(x + 14, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 14, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(11 + (row * fix2)))), min(o2,o2/(fix1/(11 + (row * fix2)))), min(o3,o3/(fix1/(11 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 15, y + 0, c1, c2, c3)
-#     canvas.SetPixel(x + 15, y + 1, c1, c2, c3)
-#     canvas.SetPixel(x + 15, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 15, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(10 + (row * fix2)))), min(o2,o2/(fix1/(10 + (row * fix2)))), min(o3,o3/(fix1/(10 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 16, y + 0, c1, c2, c3)
-#     canvas.SetPixel(x + 16, y + 1, c1, c2, c3)
-#     canvas.SetPixel(x + 16, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 16, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(9 + (row * fix2)))), min(o2,o2/(fix1/(9 + (row * fix2)))), min(o3,o3/(fix1/(9 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 17, y + 0, c1, c2, c3)
-#     canvas.SetPixel(x + 17, y + 1, c1, c2, c3)
-#     canvas.SetPixel(x + 17, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 17, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(8 + (row * fix2)))), min(o2,o2/(fix1/(8 + (row * fix2)))), min(o3,o3/(fix1/(8 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 18, y + 2, c1, c2, c3)
-#     canvas.SetPixel(x + 18, y + 3, c1, c2, c3)
-#     canvas.SetPixel(x + 18, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 18, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(7 + (row * fix2)))), min(o2,o2/(fix1/(7 + (row * fix2)))), min(o3,o3/(fix1/(7 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 19, y + 2, c1, c2, c3)
-#     canvas.SetPixel(x + 19, y + 3, c1, c2, c3)
-#     canvas.SetPixel(x + 19, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 19, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(6 + (row * fix2)))), min(o2,o2/(fix1/(6 + (row * fix2)))), min(o3,o3/(fix1/(6 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 20, y + 4, c1, c2, c3)
-#     canvas.SetPixel(x + 20, y + 5, c1, c2, c3)
-#     canvas.SetPixel(x + 20, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 20, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(5 + (row * fix2)))), min(o2,o2/(fix1/(5 + (row * fix2)))), min(o3,o3/(fix1/(5 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 21, y + 4, c1, c2, c3)
-#     canvas.SetPixel(x + 21, y + 5, c1, c2, c3)
-#     canvas.SetPixel(x + 21, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 21, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(4 + (row * fix2)))), min(o2,o2/(fix1/(4 + (row * fix2)))), min(o3,o3/(fix1/(4 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 22, y + 6, c1, c2, c3)
-#     canvas.SetPixel(x + 22, y + 7, c1, c2, c3)
-#     canvas.SetPixel(x + 22, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 22, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(3 + (row * fix2)))), min(o2,o2/(fix1/(3 + (row * fix2)))), min(o3,o3/(fix1/(3 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 23, y + 6, c1, c2, c3)
-#     canvas.SetPixel(x + 23, y + 7, c1, c2, c3)
-#     canvas.SetPixel(x + 23, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 23, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(2 + (row * fix2)))), min(o2,o2/(fix1/(2 + (row * fix2)))), min(o3,o3/(fix1/(2 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 24, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 24, y + 9, c1, c2, c3)
-#     # New line
-#     c1,c2,c3 = min(o1,o1/(fix1/(1 + (row * fix2)))), min(o2,o2/(fix1/(1 + (row * fix2)))), min(o3,o3/(fix1/(1 + (row * fix2))))
-#     print(c1,c2,c3)
-#     canvas.SetPixel(x + 25, y + 8, c1, c2, c3)
-#     canvas.SetPixel(x + 25, y + 9, c1, c2, c3)
